refactor: rename learn leftovers
