+++ packages/landingpage-new/components/getting-started/angular-guide.mdx
