+++ packages/landingpage-new/components/getting-started/javascript-guide.mdx
