+++ packages/landingpage-new/components/getting-started/layout.tsx
