+++ packages/landingpage-new/pages/getting-started/index.tsx
@@ -1,6 +1,6 @@
 import { useRouter } from 'next/router';
 import { useMount } from 'react-use';
-import { Frameworks } from '../../components/learn/frameworks';
+import { Frameworks } from '../../components/getting-started/frameworks';
 import ReactGuidePage from './react';
 
 /**
